Move updating quote attachments logic to QuoteAggregateRepository
