+++ src/VirtoCommerce.QuoteModule.ExperienceApi/Aggregates/IQuoteAggregateRepository.cs
@@ -8,4 +8,5 @@ public interface IQuoteAggregateRepository
 {
     Task<QuoteAggregate> GetById(string id);
     Task<IList<QuoteAggregate>> ToQuoteAggregates(IEnumerable<QuoteRequest> quotes);
+    Task UpdateQuoteAttachmentsAsync(QuoteRequest quote, IList<string> urls);
 }
