+++ src/VirtoCommerce.QuoteModule.ExperienceApi/Commands/DeleteQuoteAttachmentsCommandHandler.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Linq;
 using System.Threading.Tasks;
-using VirtoCommerce.FileExperienceApi.Core.Services;
 using VirtoCommerce.Platform.Core.Settings;
 using VirtoCommerce.QuoteModule.Core.Models;
 using VirtoCommerce.QuoteModule.Core.Services;
@@ -14,9 +13,8 @@ public class DeleteQuoteAttachmentsCommandHandler : UpdateQuoteAttachmentsComman
     public DeleteQuoteAttachmentsCommandHandler(
         IQuoteRequestService quoteRequestService,
         IQuoteAggregateRepository quoteAggregateRepository,
-        ISettingsManager settingsManager,
-        IFileUploadService fileUploadService)
-        : base(quoteRequestService, quoteAggregateRepository, settingsManager, fileUploadService)
+        ISettingsManager settingsManager)
+        : base(quoteRequestService, quoteAggregateRepository, settingsManager)
     {
     }
